fix: address code review — XML, session lifecycle, validation, prerequisites
- C1: Merge duplicate <devices> blocks in domain XML so vsock is inside
  the single <devices> block (libvirt rejects multiple blocks)
- C2: Add PID file management for waypipe/audio processes so stop_vm and
  destroy_vm can kill them instead of leaking via std::mem::forget
- C3+I5: Wire validate_share_path into both parse_share (CLI) and
  WayboxConfig::validate() so relative paths are rejected
- I1: Apply map_package_name to system packages before template rendering
- I6: Add socat prerequisite check
- I7: Add libvirtd connectivity check to prerequisites
- Fix clippy warnings: derive Default for NetworkMode, remove redundant
  u32 casts, fix print_literal

+++ src/libvirt/mod.rs
@@ -98,7 +98,7 @@ impl LibvirtConnection {
             let (state_raw, _) = domain.get_state()?;
             infos.push(DomainInfo {
                 name,
-                state: DomainState::from(state_raw as u32),
+                state: DomainState::from(state_raw),
             });
         }
         Ok(infos)
@@ -107,7 +107,7 @@ impl LibvirtConnection {
     pub fn get_domain_state(&self, name: &str) -> Result<DomainState> {
         let domain = self.lookup(name)?;
         let (state_raw, _) = domain.get_state()?;
-        Ok(DomainState::from(state_raw as u32))
+        Ok(DomainState::from(state_raw))
     }
 
     fn lookup(&self, name: &str) -> Result<Domain> {
